+++ base/plugins/iscore-plugin-scenario/Scenario/Document/Constraint/Rack/RackView.hpp
@@ -21,19 +21,19 @@ class RackView final : public QGraphicsObject
                            const QStyleOptionGraphicsItem* option,
                            QWidget* widget) override;
 
-        void setHeight(int height)
+        void setHeight(qreal height)
         {
             prepareGeometryChange();
             m_height = height;
         }
 
-        void setWidth(int width)
+        void setWidth(qreal width)
         {
             prepareGeometryChange();
             m_width = width;
         }
 
     private:
-        int m_height {};
-        int m_width {};
+        qreal m_height {};
+        qreal m_width {};
 };
